Remove child handling, replace with pipe signal handling.
We don't actually care about the child, really, just ability to read
and write, so SIGPIPE is more appropriate.

Currently, ignore SIGPIPE during write operations, to support just
spamming input for protocol dump testing. This should be a flag in the
future, but for now, as long as there's a match step with a thunk,
it'll get caught there.